Rename map/loading.js to MapLoading.jsx
