+++ src/js/tangram-play.js
@@ -6,7 +6,7 @@ import { tangramLayer, loadScene } from './map/map';
 import { editor, initEditor, getEditorContent, setEditorContent } from './editor/editor';
 
 // Addons
-import { showSceneLoadingIndicator, hideSceneLoadingIndicator } from './map/loading';
+import { showSceneLoadingIndicator, hideSceneLoadingIndicator } from './map/MapLoading';
 import { initWidgetMarks } from './widgets/widgets-manager';
 import { initErrorsManager } from './editor/errors';
 import { initSuggestions } from './editor/suggest';
